+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
-## [1.0.0] - 2020-05-07
+## [1.0.1] - 2020-07-06
+
+### Changed
+
+- Update the README and improve the documentation
+- Replace the JSON file to environment variable
+- Update the image result.png
+
+### Fixed
+
+- Fix the endpoint to get Cloud Function
+
+
+## [1.0.0] - 2020-07-05
 
 ### Added